Characters dpi corrected, some validations added

+++ IA-Backtracking-Using-Visual-Elements/Class/IdCost.cs
@@ -9,9 +9,9 @@ namespace IA_Backtracking_Using_Visual_Elements
     public class IdCost
     {
         public int id;
-        public int cost;
+        public float cost;
 
-        public IdCost(int id, int cost)
+        public IdCost(int id, float cost)
         {
             this.id = id;
             this.cost = cost;
